+++ src/searchForm/SearchForm.unit.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { act } from 'react-dom/test-utils'
 import { shallow, mount } from 'enzyme'
 
 import TextBody from 'typography/body'
@@ -73,6 +74,22 @@ describe('searchForm', () => {
         expect(wrapper.find(StepperOverlay).exists()).toBe(false)
         wrapper.find('.kirk-searchForm-seats > .kirk-search-button').simulate('click')
         expect(wrapper.find(StepperOverlay).exists()).toBe(true)
+      })
+
+      it('should close the overlays when clicking outiside of the form', () => {
+        const outerNode = document.createElement('div')
+        document.body.appendChild(outerNode)
+
+        wrapper = mount(<SearchForm {...defaultProps} />, {
+          attachTo: outerNode,
+        })
+
+        wrapper.find('.kirk-searchForm-seats > .kirk-search-button').simulate('click')
+        act(() => {
+          outerNode.dispatchEvent(new Event('click', { bubbles: true }))
+        })
+        wrapper.update()
+        expect(wrapper.find(StepperOverlay).exists()).toBe(false)
       })
     })